chore: support file probing on all unix systems
Allow probing of non-blockdevice backed files on all unix-like systems.
Will be used for https://github.com/siderolabs/talos/pull/11015
