+++ blkid/blkid_unix.go
@@ -2,21 +2,24 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-//go:build !unix
+//go:build unix
 
 package blkid
 
 import (
-	"fmt"
 	"os"
+
+	"golang.org/x/sys/unix"
 )
 
 // ProbePath returns the probe information for the specified path.
-func ProbePath(devpath string) (*Info, error) {
-	return nil, fmt.Errorf("not implemented")
-}
+func ProbePath(devpath string, opts ...ProbeOption) (*Info, error) {
+	f, err := os.OpenFile(devpath, os.O_RDONLY|unix.O_CLOEXEC|unix.O_NONBLOCK, 0)
+	if err != nil {
+		return nil, err
+	}
 
-// Probe returns the probe information for the specified file.
-func Probe(f *os.File) (*Info, error) {
-	return nil, fmt.Errorf("not implemented")
+	defer f.Close() //nolint:errcheck
+
+	return Probe(f, opts...)
 }
